+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -149,7 +149,8 @@ export function withExtendedResolver(
   let tsConfigResolve = tsconfig?.paths
     ? resolveWithTsConfigPaths.bind(resolveWithTsConfigPaths, {
         paths: tsconfig.paths ?? {},
-        baseUrl: tsconfig.baseUrl,
+        baseUrl: tsconfig.baseUrl ?? config.projectRoot,
+        hasBaseUrl: !!tsconfig.baseUrl,
       })
     : null;
 
@@ -170,7 +171,8 @@ export function withExtendedResolver(
             debug('Enabling tsconfig.json paths support');
             tsConfigResolve = resolveWithTsConfigPaths.bind(resolveWithTsConfigPaths, {
               paths: tsConfigPaths.paths ?? {},
-              baseUrl: tsConfigPaths.baseUrl,
+              baseUrl: tsConfigPaths.baseUrl ?? config.projectRoot,
+              hasBaseUrl: !!tsConfigPaths.baseUrl,
             });
           } else {
             debug('Disabling tsconfig.json paths support');
@@ -233,39 +235,10 @@ export function withExtendedResolver(
       );
     },
 
-    // Node.js built-ins get empty externals on web
-    (context: ResolutionContext, moduleName: string, platform: string | null) => {
-      if (
-        isFastResolverEnabled ||
-        // is web
-        platform !== 'web' ||
-        // Skip when targeting server runtimes
-        context.customResolverOptions?.environment === 'node' ||
-        // This transform only applies to Node.js built-ins
-        !isNodeExternal(moduleName)
-      ) {
-        return null;
-      }
-
-      // Perform optional resolve first. If the module doesn't exist (no module in the node_modules)
-      // then we can mock the file to use an empty module.
-      const result = getOptionalResolver(context, platform)(moduleName);
-      return (
-        result ?? {
-          // In this case, mock the file to use an empty module.
-          type: 'empty',
-        }
-      );
-    },
-
     // Node.js externals support
     (context: ResolutionContext, moduleName: string, platform: string | null) => {
-      if (
-        // is web
-        platform !== 'web' ||
-        // Only apply to server runtimes
-        context.customResolverOptions?.environment !== 'node'
-      ) {
+      // This is a web-only feature, we may extend the shimming to native platforms in the future.
+      if (platform !== 'web') {
         return null;
       }
 
@@ -273,10 +246,26 @@ export function withExtendedResolver(
       if (!moduleId) {
         return null;
       }
+
+      if (
+        // In browser runtimes, we want to either resolve a local node module by the same name, or shim the module to
+        // prevent crashing when Node.js built-ins are imported.
+        context.customResolverOptions?.environment !== 'node'
+      ) {
+        // Perform optional resolve first. If the module doesn't exist (no module in the node_modules)
+        // then we can mock the file to use an empty module.
+        const result = getOptionalResolver(context, platform)(moduleName);
+        return (
+          result ?? {
+            // In this case, mock the file to use an empty module.
+            type: 'empty',
+          }
+        );
+      }
+
       const redirectedModuleName = getNodeExternalModuleId(context.originModulePath, moduleId);
       debug(`Redirecting Node.js external "${moduleId}" to "${redirectedModuleName}"`);
-      const doResolve = getStrictResolver(context, platform);
-      return doResolve(redirectedModuleName);
+      return getStrictResolver(context, platform)(redirectedModuleName);
     },
 
     // Basic moduleId aliases
@@ -285,8 +274,7 @@ export function withExtendedResolver(
       // a way that doesn't require Babel to resolve the alias.
       if (platform && platform in aliases && aliases[platform][moduleName]) {
         const redirectedModuleName = aliases[platform][moduleName];
-        const doResolve = getStrictResolver(context, platform);
-        return doResolve(redirectedModuleName);
+        return getStrictResolver(context, platform)(redirectedModuleName);
       }
 
       return null;
@@ -351,11 +339,9 @@ export function withExtendedResolver(
       moduleName: string,
       platform: string | null
     ): CustomResolutionContext => {
-      const context = {
+      const context: Mutable<CustomResolutionContext> = {
         ...immutableContext,
-      } as Mutable<ResolutionContext> & {
-        mainFields: string[];
-        customResolverOptions?: Record<string, string>;
+        preferNativePlatform: platform !== 'web',
       };
 
       if (context.customResolverOptions?.environment === 'node') {
@@ -379,40 +365,7 @@ export function withExtendedResolver(
         }
       }
 
-      if (tsconfig?.baseUrl && isTsconfigPathsEnabled) {
-        const nodeModulesPaths: string[] = [...immutableContext.nodeModulesPaths];
-
-        if (isFastResolverEnabled) {
-          // add last to ensure node modules are resolved first
-          nodeModulesPaths.push(
-            path.isAbsolute(tsconfig.baseUrl)
-              ? tsconfig.baseUrl
-              : path.join(config.projectRoot, tsconfig.baseUrl)
-          );
-        } else {
-          // add last to ensure node modules are resolved first
-          nodeModulesPaths.push(tsconfig.baseUrl);
-        }
-
-        context.nodeModulesPaths = nodeModulesPaths;
-      }
-
-      // TODO: We can drop this in the next version upgrade (SDK 50).
-      const mainFields: string[] = context.mainFields;
-
-      return {
-        ...context,
-        preferNativePlatform: platform !== 'web',
-        // Passing `mainFields` directly won't be considered (in certain version of Metro)
-        // we need to extend the `getPackageMainPath` directly to
-        // use platform specific `mainFields`.
-        // @ts-ignore
-        getPackageMainPath(packageJsonPath) {
-          // @ts-expect-error: mainFields is not on type
-          const package_ = context.moduleCache.getPackage(packageJsonPath);
-          return package_.getMain(mainFields);
-        },
-      };
+      return context;
     }
   );